+++ packages/signers/src/lit-protocol/index.ts
@@ -1,1 +1,2 @@
 export * from "./signer.js";
+export type * from "./types.js";
